dv: fix the error setting in dv core [1/1]
PD#SWPL-3046

Problem:
movie player will keep video layer when playing video films.
Even if jumping to the next films, it will not close video layer.
When meeting this situation, dv core will keep the front
film's setting.

Solution:
when jumping the next film, we know the first frame.
So we clean the setting in the first frames.

Verify:
u212

Change-Id: I61ebc8f58baed32abc519018676e40b08d059cac
Signed-off-by: Yi Zhou <yi.zhou@amlogic.com>